+++ model.py
@@ -16,9 +16,9 @@ class Buildings(Base):
     count_floors = Column(Integer)
     sections = relationship('Sections', back_populates = 'building')
     photo = Column(String)
-    
+       
     def __repr__(self):
-        return f"{self.street}, {self.index}"
+        return f"{self.sections}, {self.index}"
 
 
 class Tenants(Base):
