--- models/group.py
@@ -1,14 +1,0 @@
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.orm import relationship
-
-from models.database import Base
-
-
-class Group(Base):
-    __tablename__ = 'groups'
-
-    id = Column(Integer, primary_key=True)
-    group_name = Column(String)
-    student = relationship('Student')
-
-
