+++ webapp/templates/show_buldings.html
@@ -1,0 +1,11 @@
+<div class="containermt-5">
+  <h1> Промышленные здания </h1>
+  {% for element in building %}
+  <div class="alert alert-info">
+      <h2> {{ element.street }} - {{ element.index }} </h2>
+      <p> Этажей: {{ element.count_floors }}</p>
+      <a href="/sections" class="btn btn-warning">Подробнее</a>
+  </div>
+  {% endfor %}
+</div>
+
