+++ webapp/templates/show_sections.html
@@ -1,0 +1,10 @@
+<div class="containermt-5">
+  <h1> Промышленные здания </h1>
+  {% for id in sectionss.id %}
+        while {{id == buildings_id}}
+            <h2> {{ sectionss.floor }} </h2>
+            <p> Этажей: </p>
+        
+  {% endfor %}
+</div>
+
